Rename Alluxio worker classes
Rename interface AlluxioWorker to AlluxioWorkerService
Rename main() class AlluxioWorkerLauncher to AlluxioWorker
