+++ bin/alluxio-start.sh
@@ -168,7 +168,7 @@ start_worker() {
   echo "Starting worker @ $(hostname -f). Logging to ${ALLUXIO_LOGS_DIR}"
   (nohup ${JAVA} -cp ${CLASSPATH} \
    ${ALLUXIO_WORKER_JAVA_OPTS} \
-   alluxio.worker.AlluxioWorkerLauncher > ${ALLUXIO_LOGS_DIR}/worker.out 2>&1 ) &
+   alluxio.worker.AlluxioWorker > ${ALLUXIO_LOGS_DIR}/worker.out 2>&1 ) &
 }
 
 restart_worker() {
@@ -176,12 +176,12 @@ restart_worker() {
     ALLUXIO_WORKER_JAVA_OPTS=${ALLUXIO_JAVA_OPTS}
   fi
 
-  RUN=$(ps -ef | grep "alluxio.worker.AlluxioWorkerLauncher" | grep "java" | wc | cut -d" " -f7)
+  RUN=$(ps -ef | grep "alluxio.worker.AlluxioWorker" | grep "java" | wc | cut -d" " -f7)
   if [[ ${RUN} -eq 0 ]]; then
     echo "Restarting worker @ $(hostname -f). Logging to ${ALLUXIO_LOGS_DIR}"
     (nohup ${JAVA} -cp ${CLASSPATH} \
      ${ALLUXIO_WORKER_JAVA_OPTS} \
-     alluxio.worker.AlluxioWorkerLauncher > ${ALLUXIO_LOGS_DIR}/worker.out 2>&1) &
+     alluxio.worker.AlluxioWorker > ${ALLUXIO_LOGS_DIR}/worker.out 2>&1) &
   fi
 }
