+++ core/server/src/main/java/alluxio/worker/WorkerIdRegistry.java
@@ -27,7 +27,7 @@ import javax.annotation.concurrent.NotThreadSafe;
 /**
  * The single place to get, set, and update worker id.
  *
- * When worker process starts in {@link AlluxioWorker}, worker is temporarily lost from master,
+ * When worker process starts in {@link AlluxioWorkerService}, worker is temporarily lost from master,
  * or when leader master is down in fault tolerant mode, this class will try to get a new worker id
  * from {@link BlockMaster}.
  *
