+++ core/server/src/test/java/alluxio/worker/block/SpaceReserverTest.java
@@ -14,7 +14,7 @@ package alluxio.worker.block;
 import alluxio.Configuration;
 import alluxio.ConfigurationTestUtils;
 import alluxio.Constants;
-import alluxio.worker.AlluxioWorker;
+import alluxio.worker.AlluxioWorkerService;
 import alluxio.worker.DataServer;
 
 import org.junit.After;
@@ -76,7 +76,7 @@ public class SpaceReserverTest {
     // We shouldn't need to bind net addresses in unit tests
     PowerMockito.mockStatic(DataServer.Factory.class);
     PowerMockito.when(
-        DataServer.Factory.create(Mockito.<InetSocketAddress>any(), Mockito.<AlluxioWorker>any()))
+        DataServer.Factory.create(Mockito.<InetSocketAddress>any(), Mockito.<AlluxioWorkerService>any()))
         .thenReturn(Mockito.mock(DataServer.class));
 
     BlockWorker blockWorker = new BlockWorker();
